requirements.txt
accelerate==1.10.0
aiofiles==24.1.0
alembic==1.16.4
autogen==0.9.9
docling==2.47.0
fastapi==0.115.14
langchain==1.2.3
langchain-community==0.4.1
langchain-huggingface==1.2.0
langchain-ollama==1.0.1
langchain_text_splitters==1.1.0
loguru==0.7.3
openai==2.15.0
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
python-multipart==0.0.20
qdrant-client==1.15.1
transformers==4.55.4
sentence-transformers==5.2.0
scikit_learn==1.8.0
uvicorn==0.35.0